Change the C types to [c_]int for simplicity...

+++ test/extern/rename/ccode.h
@@ -3,10 +3,10 @@
 int64_t type = 42;
 
 typedef struct _S {
-  int64_t okname;
-  int64_t type;
+  int okname;
+  int type;
 } S;
 
 static inline void printS(S* myS) {
-  printf("{ %lld %lld}\n", myS->okname, myS->type);
+  printf("{ %d %d}\n", myS->okname, myS->type);
 }
